Revert "Rename _slater.py to slater.py"
This reverts commit 88c19b6365fc30a74c388bd851b225d36207ba9a.
