+++ bfit/_slater.py
